+++ lib/oxaiworkers/assistant/coder.rb
@@ -7,9 +7,18 @@ module OxAiWorkers
 
       def initialize(delayed: false, model: nil, language: 'ruby')
         store_locale
+
+        with_locale do
+          @id = 'coder'
+          @role = format(I18n.t('oxaiworkers.assistant.coder.role'), language)
+          @description = I18n.t('oxaiworkers.assistant.coder.description')
+          @capabilities = I18n.t('oxaiworkers.assistant.coder.capabilities').split(',').map(&:strip)
+          @title = I18n.t('oxaiworkers.assistant.coder.title')
+        end
+
         @iterator = Iterator.new(
           worker: init_worker(delayed:, model:),
-          role: format(I18n.t('oxaiworkers.assistant.coder.role'), language),
+          role: @role,
           tools: [Tool::Eval.new, Tool::FileSystem.new],
           locale: @locale,
           on_inner_monologue: ->(text:) { puts "monologue: #{text}".colorize(:yellow) },
@@ -19,7 +28,8 @@ module OxAiWorkers
 
       def language=(language)
         with_locale do
-          @iterator.role = format(I18n.t('oxaiworkers.assistant.coder.role'), language)
+          @role = format(I18n.t('oxaiworkers.assistant.coder.role'), language)
+          @iterator.role = @role
         end
       end
     end