+++ lib/oxaiworkers/assistant/localizer.rb
@@ -7,9 +7,18 @@ module OxAiWorkers
 
       def initialize(delayed: false, model: nil, language: 'русский', locale: :ru, source: 'english')
         store_locale
+
+        with_locale do
+          @id = 'localizer'
+          @role = format(I18n.t('oxaiworkers.assistant.localizer.role'), language)
+          @description = I18n.t('oxaiworkers.assistant.localizer.description')
+          @capabilities = I18n.t('oxaiworkers.assistant.localizer.capabilities').split(',').map(&:strip)
+          @title = I18n.t('oxaiworkers.assistant.localizer.title')
+        end
+
         @iterator = Iterator.new(
           worker: init_worker(delayed:, model:),
-          role: format(I18n.t('oxaiworkers.assistant.localizer.role'), language),
+          role: @role,
           tools: [Tool::Eval.new, Tool::FileSystem.new],
           locale: @locale,
           on_inner_monologue: ->(text:) { puts "monologue: #{text}".colorize(:yellow) },