+++ lib/oxaiworkers/assistant/sysop.rb
@@ -7,9 +7,18 @@ module OxAiWorkers
 
       def initialize(delayed: false, model: nil)
         store_locale
+
+        with_locale do
+          @id = 'sysop'
+          @role = I18n.t('oxaiworkers.assistant.sysop.role')
+          @description = I18n.t('oxaiworkers.assistant.sysop.description')
+          @capabilities = I18n.t('oxaiworkers.assistant.sysop.capabilities').split(',').map(&:strip)
+          @title = I18n.t('oxaiworkers.assistant.sysop.title')
+        end
+
         @iterator = Iterator.new(
           worker: init_worker(delayed:, model:),
-          role: I18n.t('oxaiworkers.assistant.sysop.role'),
+          role: @role,
           tools: [Tool::Eval.new(only: :sh), Tool::FileSystem.new(only: %i[read_file write_to_file])],
           locale: @locale,
           on_inner_monologue: ->(text:) { puts "monologue: #{text}".colorize(:yellow) },